min->std::min max->std::max
git-svn-id: svn://fileserver/activex/AVS/Sources/TeamlabOffice/trunk/ServerComponents@59184 954022d7-b5bf-4e40-9824-e11837661b57

+++ ASCPresentationEditor/OfficeDrawing/Shapes/BaseShape/PPTXShape/Formula.h
@@ -340,14 +340,14 @@ namespace NSGuidesOOXML
 		{
 			//for(long i = 0; i < Adjustments.size(); i++)
 			//	Adjustments[i] = NonDefResult;
-			for(long i = 0; i < Guides->size(); i++)
+            for(size_t i = 0; i < Guides->size(); i++)
 				(*Guides)[i] = dNonDefResult;
 		}
 
 		void ReCalculateGuides()
 		{
 			Clear();
-			for(long i = 0; i < strGuides.size(); i++)
+            for(size_t i = 0; i <  strGuides.size(); i++)
 				(*Guides)[i] = strGuides[i].Calculate(this);
 		}
 	};
